% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gpas}
\alias{gpas}
\title{Helper function to perform the (de)pseudonymization}
\usage{
gpas(
  GPAS_BASE_URL = NULL,
  GPAS_PSEUDONYM_DOMAIN = NULL,
  depseudonymize = FALSE,
  allow_create = TRUE,
  gpas_fieldvalue,
  from_env = FALSE,
  handle_na = TRUE
)
}
\arguments{
\item{GPAS_BASE_URL}{(Optional, String)
The URL to your gPAS API.
E.g. 'https://gpas.hospital.org'.}

\item{GPAS_PSEUDONYM_DOMAIN}{(Optional, String) The name of
the pseudonym domain configured in gPAS.}

\item{depseudonymize}{(boolean, default = FALSE).
Do you want to depseudonymize the values? If false, the values will be
pseudonymized.}

\item{allow_create}{(boolean, default = TRUE). Do you want to create
new pseudonyms if there is no pseudonym for some of your values yet?
The value of this parameter has no effect, if `depseudonymize == TRUE`.}

\item{gpas_fieldvalue}{(String) The actual value(s) to pseudonymized.}

\item{from_env}{(Optional, Boolean, Default = `TRUE`) If true, the
connection parameters `GPAS_BASE_URL`, `GPAS_PSEUDONYM_DOMAIN`
are read from the environment
and can therefore be left empty when calling this function.}
}
\value{
(vector) All pseudonyms for the input values.
}
\description{
Helper function to perform the (de)pseudonymization
}
