% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depseudonymize.R
\name{depseudonymize}
\alias{depseudonymize}
\title{Convert a pseudonym back to its input value.}
\usage{
depseudonymize(
  GPAS_BASE_URL = NULL,
  GPAS_API_KEY = NULL,
  GPAS_FIELDNAME = NULL,
  gpas_fieldvalue,
  from_env = FALSE
)
}
\arguments{
\item{GPAS_BASE_URL}{(Optional, String)
The URL to your gPAS API.
E.g. 'https://ml.hospital.de'.}

\item{GPAS_API_KEY}{(Optional, String)
The API key which is allowed to access the gPAS.}

\item{GPAS_FIELDNAME}{(Optional, String) The name of
the field to use for the gPAS. Specified in the ML-config.}

\item{gpas_fieldvalue}{(String) The actual value to de-pseudonymized.}

\item{from_env}{(Optional, Boolean, Default = `FALSE`) If true, the
connection parameters `GPAS_BASE_URL`, `GPAS_API_KEY` and
`GPAS_FIELDNAME` are read from the environment
and can therefore be left empty when calling this function.}
}
\value{
(vector) All pseudonyms for the input values.
}
\description{
Convert a pseudonym back to its input value.
}
