% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonymize.R
\name{pseudonymize}
\alias{pseudonymize}
\title{Get a pseudonym for a value}
\usage{
pseudonymize(
  GPAS_BASE_URL = NULL,
  GPAS_PSEUDONYM_DOMAIN = NULL,
  allow_create = TRUE,
  gpas_fieldvalue,
  from_env = FALSE
)
}
\arguments{
\item{GPAS_BASE_URL}{(Optional, String)
The URL to your gPAS API.
E.g. 'https://gpas.hospital.org'.}

\item{GPAS_PSEUDONYM_DOMAIN}{(Optional, String) The name of
the pseudonym domain configured in gPAS.}

\item{allow_create}{(boolean, default = TRUE). Do you want to create
new pseudonyms if there is no pseudonym for some of your values yet?
The value of this parameter has no effect, if `depseudonymize == TRUE`.}

\item{gpas_fieldvalue}{(String) The actual value(s) to pseudonymized.}

\item{from_env}{(Optional, Boolean, Default = `FALSE`) If true, the
connection parameters `GPAS_BASE_URL`, `GPAS_PSEUDONYM_DOMAIN`
are read from the environment
and can therefore be left empty when calling this function.}
}
\value{
(vector) All pseudonyms for the input values.
}
\description{
Get a pseudonym for a value
}
